// apps/trackflix/src/API.ts
export enum Source {
  SELF = "SELF",
  YOUTUBE = "YOUTUBE",
  LIVESTREAM_SELF = "LIVESTREAM_SELF",
}
// ...
export type Media = {
  __typename: "Media",
  id: string,
  title: string,
  description: string,
  highlighted: boolean,
  source?: Source | null,
  author: string,
  viewCount?: number | null,
  createdAt: string,
  updatedAt: string,
  thumbnail?: Thumbnail | null,
  sections?: ModelMediasSectionsConnection | null,
};

export type Thumbnail = {
  __typename: "Thumbnail",
  id: string,
  ext: string,
  src?: string | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelMediasSectionsConnection = {
  __typename: "ModelMediasSectionsConnection",
  items:  Array<MediasSections | null >,
  nextToken?: string | null,
};
// ...
export type SecuredLink = {
  __typename: "SecuredLink",
  id: string,
  mediaID: string,
  usagesRemaining: number,
  notifiedEmail?: string | null,
  expirationUnixTime: number,
  createdAt: string,
  updatedAt: string,
  media: Media,
};
// ...
export type SecuredLinkLog = {
  __typename: "SecuredLinkLog",
  id: string,
  mediaID: string,
  securedLinkId: string,
  message?: string | null,
  createdAt: string,
  updatedAt: string,
  media: Media,
  securedLink?: SecuredLink | null,
};